+++ src/rpc/server.cpp
@@ -88,10 +88,10 @@ void RPCServerSignals::OnStopped(std::function<void()> slot) {
 }
 
 void RPCTypeCheck(const UniValue &params,
-                  const std::list<UniValue::VType> &typesExpected,
+                  const std::list<UniValueType> &typesExpected,
                   bool fAllowNull) {
     unsigned int i = 0;
-    for (UniValue::VType t : typesExpected) {
+    for (const UniValueType &t : typesExpected) {
         if (params.size() <= i) {
             break;
         }
@@ -104,11 +104,13 @@ void RPCTypeCheck(const UniValue &params,
     }
 }
 
-void RPCTypeCheckArgument(const UniValue &value, UniValue::VType typeExpected) {
-    if (value.type() != typeExpected) {
-        throw JSONRPCError(RPC_TYPE_ERROR, strprintf("Expected type %s, got %s",
-                                                     uvTypeName(typeExpected),
-                                                     uvTypeName(value.type())));
+void RPCTypeCheckArgument(const UniValue &value,
+                          const UniValueType &typeExpected) {
+    if (!typeExpected.typeAny && value.type() != typeExpected.type) {
+        throw JSONRPCError(RPC_TYPE_ERROR,
+                           strprintf("Expected type %s, got %s",
+                                     uvTypeName(typeExpected.type),
+                                     uvTypeName(value.type())));
     }
 }