+++ src/rpc/server.h
@@ -39,11 +39,10 @@ class CNetAddr;
 
 /**
  * Wrapper for UniValue::VType, which includes typeAny: used to denote don't
- * care type. Only used by RPCTypeCheckObj.
+ * care type.
  */
 struct UniValueType {
-    explicit UniValueType(UniValue::VType _type)
-        : typeAny(false), type(_type) {}
+    UniValueType(UniValue::VType _type) : typeAny(false), type(_type) {}
     UniValueType() : typeAny(true) {}
     bool typeAny;
     UniValue::VType type;
@@ -101,13 +100,14 @@ bool RPCIsInWarmup(std::string *outStatus);
  * correct type.
  */
 void RPCTypeCheck(const UniValue &params,
-                  const std::list<UniValue::VType> &typesExpected,
+                  const std::list<UniValueType> &typesExpected,
                   bool fAllowNull = false);
 
 /**
  * Type-check one argument; throws JSONRPCError if wrong type given.
  */
-void RPCTypeCheckArgument(const UniValue &value, UniValue::VType typeExpected);
+void RPCTypeCheckArgument(const UniValue &value,
+                          const UniValueType &typeExpected);
 
 /**
  * Check for expected keys/value types in an Object.