+++ src/test/rpc_tests.cpp
@@ -53,8 +53,6 @@ BOOST_AUTO_TEST_CASE(rpc_rawparams) {
                       std::runtime_error);
     BOOST_CHECK_THROW(CallRPC("createrawtransaction not_array"),
                       std::runtime_error);
-    BOOST_CHECK_THROW(CallRPC("createrawtransaction [] []"),
-                      std::runtime_error);
     BOOST_CHECK_THROW(CallRPC("createrawtransaction {} {}"),
                       std::runtime_error);
     BOOST_CHECK_NO_THROW(CallRPC("createrawtransaction [] {}"));